Fix LLVM warnings and asssociated minor fixes (#93)
This PR fixes warnings when compiling with Apple clang. This (and the tests) revealed a few issues, that were fixed when possible. In particular:

Bug fixes:

    There were a couple bugs in the FSI solver, the main one being that with the local master to local accumulator coupling, the ghosted force accumulators were only added to ranks which themself owned an accumulator. But, there can be more ranks with position masters than ranks with accumulators, the reason being that the former is tied to node-based dofs, whereas the latter are defined based on owned faces on the cylinder, as they compute integrals. This hopefully has been fixed, and the accumulators are ghosted on all ranks that have position masters.
    With nonzero mass, invalid cylinder velocity dofs were added to the sparsity pattern on ranks where these dofs were not set. I don't really understand why this did not show up in the debug tests before.
    The incompressible NS solver was not actually using its coupling table to create the sparsity pattern.

Some issues were revealed because Metis' partitioning is likely not deterministic, e.g., I had a problematic partition on MacOS with no face on the cylinder but still some owned dofs on it, while I had a different partitioning on Ubuntu, for which the tests passed.

Warning fixes:

    The warnings from fez have been corrected
    The warnings originating from Eigen have been silenced by flagging Eigen files as SYSTEM includes.

+++ src/error_estimation/solution_recovery.cpp
@@ -108,10 +108,11 @@ namespace ErrorEstimation
       n_fields_to_recover = 1;
       for (unsigned int i = 1; i <= degree; ++i)
       {
-        n_fields_to_recover += std::pow(dim, i);
+        n_fields_to_recover += static_cast<unsigned int>(std::pow(dim, i));
       }
       n_derivatives_to_store =
-        n_fields_to_recover - 1 + std::pow(dim, degree + 1);
+        n_fields_to_recover - 1 +
+        static_cast<unsigned int>(std::pow(dim, degree + 1));
     }
 
     template <int dim>
@@ -739,8 +740,10 @@ namespace ErrorEstimation
 
       // Total number of vector components in this object's FESystem
       // = 1 + dim + dim^2 + ... = (dim^(N+1) - 1) / (dim - 1)
-      this->n_components =
-        (std::pow(dim, highest_recovered_derivative + 1) - 1) / (dim - 1);
+      this->n_components = (static_cast<unsigned int>(
+                              std::pow(dim, highest_recovered_derivative + 1)) -
+                            1) /
+                           (dim - 1);
 
       const auto comm     = this->mpi_communicator;
       auto      &owned    = this->locally_owned_recovery_dofs;
